# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/templates": {
      "get": {
        "operationId": "TemplatesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "example": "เกียรติบัตร",
              "type": "string"
            }
          },
          {
            "name": "designType",
            "required": false,
            "in": "query",
            "schema": {
              "example": "certificate",
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "example": "education",
              "type": "string"
            }
          },
          {
            "name": "owner",
            "required": false,
            "in": "query",
            "description": "me = ที่ฉันเผยแพร่ · others = ของคนอื่น",
            "schema": {
              "type": "string",
              "enum": [
                "me",
                "others"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "popular",
              "type": "string",
              "enum": [
                "popular",
                "recent"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TemplateSummaryDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เทมเพลตทั้งหมด (ยอดนิยม · ล่าสุด · กรองประเภท/หมวด/เจ้าของ)",
        "tags": [
          "templates"
        ]
      },
      "post": {
        "operationId": "TemplatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/TemplateDto"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "นักศึกษาและผู้เยี่ยมชมเผยแพร่เทมเพลตไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "NOT_FOUND"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เผยแพร่งานเป็นเทมเพลต (อาจารย์ บุคลากร และผู้ดูแลเท่านั้น)",
        "tags": [
          "templates"
        ]
      }
    },
    "/api/v1/templates/{id}": {
      "get": {
        "operationId": "TemplatesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/TemplateDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เทมเพลตหนึ่งชิ้นพร้อม JSON state",
        "tags": [
          "templates"
        ]
      },
      "patch": {
        "operationId": "TemplatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/TemplateDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "แก้ชื่อ คำอธิบาย หรือหมวดของเทมเพลต",
        "tags": [
          "templates"
        ]
      },
      "delete": {
        "operationId": "TemplatesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เลิกเผยแพร่เทมเพลต",
        "tags": [
          "templates"
        ]
      }
    },
...
  },
  "info": {
    "title": "CS Canvas — csmju-canvas",
    "description": "ระบบย่อยภายใต้ CSMJU2030 · ตัวตนมาจาก JWT ของ Core Hub ที่ระบบนี้ตรวจลายเซ็นเองด้วย JWKS ทุก request (RS256 · iss=core-hub · aud=csmju2030 · kid=core-hub-2026) ส่ง token มาได้สองทาง: Authorization: Bearer หรือคุกกี้ csmju_canvas_access_token (ตั้งโดย /auth/callback)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csmju_canvas_access_token"
      }
    },
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 2480
          },
          "page": {
            "type": "number",
            "example": 2
          },
          "limit": {
            "type": "number",
            "example": 50
          },
          "totalPages": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "DeletedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
...
      "TemplateSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "designType": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "thumbnail": {
            "type": "object",
            "nullable": true
          },
          "usageCount": {
            "type": "number"
          },
          "isBuiltIn": {
            "type": "boolean",
            "description": "true = เทมเพลตตั้งต้นของทีม CS Canvas"
          },
          "isMine": {
            "type": "boolean",
            "description": "ผู้เรียกเป็นคนเผยแพร่เอง"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "designType",
          "category",
          "width",
          "height",
          "usageCount",
          "isBuiltIn",
          "isMine",
          "createdAt",
          "updatedAt"
        ]
      },
      "TemplateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "designType": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "thumbnail": {
            "type": "object",
            "nullable": true
          },
          "usageCount": {
            "type": "number"
          },
          "isBuiltIn": {
            "type": "boolean",
            "description": "true = เทมเพลตตั้งต้นของทีม CS Canvas"
          },
          "isMine": {
            "type": "boolean",
            "description": "ผู้เรียกเป็นคนเผยแพร่เอง"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "document": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "designType",
          "category",
          "width",
          "height",
          "usageCount",
          "isBuiltIn",
          "isMine",
          "createdAt",
          "updatedAt",
          "document"
        ]
      },
      "CreateTemplateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "เกียรติบัตรงานแข่งขันโปรแกรม"
          },
          "description": {
            "type": "string",
            "example": "แก้ชื่อผู้รับและวันที่ได้ทันที"
          },
          "designType": {
            "type": "string",
            "example": "certificate"
          },
          "category": {
            "type": "string",
            "example": "education"
          },
          "width": {
            "type": "number",
            "example": 1754
          },
          "height": {
            "type": "number",
            "example": 1240
          },
          "document": {
            "type": "object",
            "additionalProperties": true
          },
          "thumbnail": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "designType",
          "category",
          "width",
          "height",
          "document"
        ]
      },
      "UpdateTemplateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          }
        }
      },
...
    }
  }
}